% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_blocks.R
\name{add_change_from_baseline}
\alias{add_change_from_baseline}
\title{Add change from baseline values:
 Value, Change, Percent Change from baseline}
\usage{
add_change_from_baseline(data, df_name = "LB")
}
\arguments{
\item{df_name}{prefix of domain variables - name of SDTM}
}
\value{
dataframe
}
\description{
Add change from baseline values:
 Value, Change, Percent Change from baseline
}
