% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_blocks.R
\name{lbl}
\alias{lbl}
\title{Create a labeled variable}
\usage{
lbl(var, label)
}
\arguments{
\item{var}{The variable to assign the label to}

\item{label}{The label to assign to the variable.}
}
\value{
A labeled variable with the specified label.
}
\description{
This function creates a labeled variable by attaching a label attribute to a given value.
}
\examples{
lbl(iris$Species, "Flower Species")

}
